feat: incluir president na lista de leitor
