+++ app/src/main/java/br/com/bvilela/listbuilder/config/NotifyProperties.java
@@ -3,12 +3,11 @@ package br.com.bvilela.listbuilder.config;
 import br.com.bvilela.listbuilder.enuns.DayOfWeekEnum;
 import br.com.bvilela.listbuilder.exception.ListBuilderException;
 import br.com.bvilela.listbuilder.utils.AppUtils;
+import java.util.List;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-
-import java.util.List;
 
 @Getter
 @Component
