+++ app/src/main/java/br/com/bvilela/listbuilder/dto/discourse/writer/DiscourseWriterDTO.java
@@ -1,0 +1,16 @@
+package br.com.bvilela.listbuilder.dto.discourse.writer;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DiscourseWriterDTO {
+    private List<DiscourseWriterItemDTO> send;
+    private List<DiscourseWriterItemDTO> receive;
+}
