+++ app/src/main/java/br/com/bvilela/listbuilder/service/impl/GroupServiceImpl.java
@@ -1,9 +1,9 @@
 package br.com.bvilela.listbuilder.service.impl;
 
 import br.com.bvilela.listbuilder.config.MessageConfig;
+import br.com.bvilela.listbuilder.dto.InputListDTO;
 import br.com.bvilela.listbuilder.dto.ItemDateDTO;
 import br.com.bvilela.listbuilder.dto.designacao.FileInputDataDesignacaoDTO;
-import br.com.bvilela.listbuilder.dto.InputListDTO;
 import br.com.bvilela.listbuilder.dto.designacao.writer.DesignacaoWriterItemDTO;
 import br.com.bvilela.listbuilder.dto.limpeza.FileInputDataLimpezaDTO;
 import br.com.bvilela.listbuilder.enuns.DesignacaoEntityEnum;
@@ -161,9 +161,7 @@ public class GroupServiceImpl implements GroupService {
 
     @SneakyThrows
     private List<DesignacaoWriterItemDTO> generateSequenceListDesignacao(
-            InputListDTO dto,
-            List<LocalDate> listDates,
-            DesignacaoEntityEnum entity) {
+            InputListDTO dto, List<LocalDate> listDates, DesignacaoEntityEnum entity) {
 
         var indexLast = dto.getList().indexOf(dto.getLast());
         if (indexLast < 0) {
