+++ app/src/main/resources/application.properties
@@ -18,6 +18,7 @@ layout.limpeza=2
 
 ### Config Designacoes ###
 designacao.gerar.automatico.ind.mic=true
+notify.designation.type.active=Leitor
 
 ### Notification ###
 notif.active=true
