+++ app/src/test/java/br/com/bvilela/listbuilder/builder/designacao/FileInputDataDesignacaoDtoBuilder.java
@@ -1,7 +1,7 @@
 package br.com.bvilela.listbuilder.builder.designacao;
 
-import br.com.bvilela.listbuilder.dto.designacao.FileInputDataDesignacaoDTO;
 import br.com.bvilela.listbuilder.dto.InputListDTO;
+import br.com.bvilela.listbuilder.dto.designacao.FileInputDataDesignacaoDTO;
 import br.com.bvilela.listbuilder.dto.designacao.FileInputDataDesignacaoReaderDTO;
 import java.util.List;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -60,8 +60,7 @@ public class FileInputDataDesignacaoDtoBuilder {
         return this;
     }
 
-    public FileInputDataDesignacaoDtoBuilder withPresident(
-            InputListDTO president) {
+    public FileInputDataDesignacaoDtoBuilder withPresident(InputListDTO president) {
         this.target.setPresident(president);
         return this;
     }
@@ -71,8 +70,7 @@ public class FileInputDataDesignacaoDtoBuilder {
         return this;
     }
 
-    public FileInputDataDesignacaoDtoBuilder withAudioVideo(
-            InputListDTO audioVideo) {
+    public FileInputDataDesignacaoDtoBuilder withAudioVideo(InputListDTO audioVideo) {
         this.target.setAudioVideo(audioVideo);
         return this;
     }
