+++ app/src/test/java/br/com/bvilela/listbuilder/service/audience/impl/AudienceWriterServiceImplTest.java
@@ -1,20 +1,19 @@
 package br.com.bvilela.listbuilder.service.audience.impl;
 
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
 import br.com.bvilela.listbuilder.config.AppProperties;
 import br.com.bvilela.listbuilder.enuns.AudienceWriterLayoutEnum;
 import br.com.bvilela.listbuilder.utils.PropertiesTestUtils;
 import br.com.bvilela.listbuilder.utils.TestUtils;
+import java.nio.file.Paths;
+import java.util.List;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.MockitoAnnotations;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.nio.file.Paths;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 @SpringBootApplication
 class AudienceWriterServiceImplTest {
