+++ app/src/test/java/br/com/bvilela/listbuilder/service/impl/GroupServiceImplTest.java
@@ -4,8 +4,8 @@ import br.com.bvilela.listbuilder.builder.FileInputDataLimpezaDtoBuilder;
 import br.com.bvilela.listbuilder.builder.designacao.DesignacaoWriterItemDtoBuilder;
 import br.com.bvilela.listbuilder.builder.designacao.FileInputDataDesignacaoDtoBuilder;
 import br.com.bvilela.listbuilder.config.MessageConfig;
-import br.com.bvilela.listbuilder.dto.ItemDateDTO;
 import br.com.bvilela.listbuilder.dto.InputListDTO;
+import br.com.bvilela.listbuilder.dto.ItemDateDTO;
 import br.com.bvilela.listbuilder.dto.designacao.writer.DesignacaoWriterItemDTO;
 import br.com.bvilela.listbuilder.enuns.DesignacaoEntityEnum;
 import br.com.bvilela.listbuilder.exception.ListBuilderException;
@@ -281,8 +281,7 @@ class GroupServiceImplTest {
     }
 
     //	 --------------------- OTHER --------------------- \\
-    private void validListGenerated(
-            InputListDTO dto, List<DesignacaoWriterItemDTO> list) {
+    private void validListGenerated(InputListDTO dto, List<DesignacaoWriterItemDTO> list) {
         Assertions.assertEquals(4, list.size());
         Assertions.assertEquals(dto.getList().get(1), list.get(0).getName());
         Assertions.assertEquals(dto.getList().get(2), list.get(1).getName());
